Make TableConfigTunerRegistry configurable to scan packages. (#7089)
