+++ pinot-controller/src/main/java/org/apache/pinot/controller/tuner/TableConfigTuner.java
@@ -16,23 +16,25 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.config.tuner;
+package org.apache.pinot.controller.tuner;
 
 import org.apache.pinot.spi.config.table.TableConfig;
 import org.apache.pinot.spi.config.table.TunerConfig;
-import org.apache.pinot.spi.config.table.tuner.TableConfigTuner;
-import org.apache.pinot.spi.config.table.tuner.Tuner;
 import org.apache.pinot.spi.data.Schema;
 
 
-@Tuner(name = "noopConfigTuner")
-public class NoOpTableTableConfigTuner implements TableConfigTuner {
-  @Override
-  public void init(TunerConfig props, Schema schema) {
-  }
+/**
+ * Interface for Table Config Tuner.
+ */
+public interface TableConfigTuner {
+  /**
+   * Used to initialize underlying implementation with Schema
+   * and custom properties (eg: metrics end point)
+   */
+  void init(TunerConfig props, Schema schema);
 
-  @Override
-  public TableConfig apply(TableConfig initialConfig) {
-    return initialConfig;
-  }
+  /**
+   * Takes the original TableConfig and returns a tuned one
+   */
+  TableConfig apply(TableConfig initialConfig);
 }
