+++ pinot-controller/src/main/java/org/apache/pinot/controller/tuner/Tuner.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.spi.config.table.tuner;
+package org.apache.pinot.controller.tuner;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
